requirements.txt
numpy<=1.19.5
scikit-learn>=0.24.2
scipy>=1.5.4
opencv-python>=4.5.2.54
tensorflow>=2.6.2
python-mnist>=0.7
